Move stale output cleanup to execution engine
