+++ subprojects/integ-test/src/integTest/groovy/org/gradle/integtests/StaleOutputIntegrationTest.groovy
@@ -16,9 +16,9 @@
 
 package org.gradle.integtests
 
-import org.gradle.api.internal.tasks.execution.CleanupStaleOutputsExecuter
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 import org.gradle.integtests.fixtures.BuildOperationsFixture
+import org.gradle.internal.execution.steps.CleanupStaleOutputsStep
 import org.gradle.test.fixtures.file.TestFile
 import org.gradle.util.ToBeImplemented
 import spock.lang.Issue
@@ -123,7 +123,7 @@ class StaleOutputIntegrationTest extends AbstractIntegrationSpec {
     def "two tasks can output in the same directory with --rerun-tasks"() {
         buildFile << """
             apply plugin: 'base'
-            
+
             task firstCopy {
                 inputs.file('first.file')
                 outputs.dir('build/destination')
@@ -131,7 +131,7 @@ class StaleOutputIntegrationTest extends AbstractIntegrationSpec {
                     file('build/destination/first.file').text = file('first.file').text
                 }
             }
-            
+
             task secondCopy {
                 inputs.file('second.file')
                 outputs.dir('build/destination')
@@ -139,7 +139,7 @@ class StaleOutputIntegrationTest extends AbstractIntegrationSpec {
                     file('build/destination/second.file').text = file('second.file').text
                 }
             }
-            
+
             secondCopy.dependsOn firstCopy
         """
         file("first.file").createFile()
@@ -156,14 +156,14 @@ class StaleOutputIntegrationTest extends AbstractIntegrationSpec {
         given:
         buildFile << """
             apply plugin: 'base'
-            
+
             task myTask {
                 outputs.dir "external/output"
                 outputs.file "customFile"
                 outputs.dir "build/dir"
                 doLast {}
             }
-            
+
             clean {
                 delete "customFile"
             }
@@ -291,7 +291,7 @@ class StaleOutputIntegrationTest extends AbstractIntegrationSpec {
 
         then:
         fixture.staleFilesHaveBeenRemoved()
-        operations.hasOperation(CleanupStaleOutputsExecuter.CLEAN_STALE_OUTPUTS_DISPLAY_NAME)
+        operations.hasOperation(CleanupStaleOutputsStep.CLEAN_STALE_OUTPUTS_DISPLAY_NAME)
     }
 
     def "no build operations are created for stale outputs cleanup if no files are removed"() {
@@ -305,7 +305,7 @@ class StaleOutputIntegrationTest extends AbstractIntegrationSpec {
 
         then:
         executedAndNotSkipped(fixture.taskPath)
-        !operations.first(CleanupStaleOutputsExecuter.CLEAN_STALE_OUTPUTS_DISPLAY_NAME)
+        !operations.first(CleanupStaleOutputsStep.CLEAN_STALE_OUTPUTS_DISPLAY_NAME)
     }
 
     def "overlapping outputs between 'build/outputs' and '#overlappingOutputDir' are not cleaned up"() {
@@ -363,7 +363,7 @@ class StaleOutputIntegrationTest extends AbstractIntegrationSpec {
         String taskName = 'test'
 
         String getBuildScript() {
-            """       
+            """
                 apply plugin: 'base'
 
                 task ${taskName} {
@@ -517,14 +517,14 @@ class StaleOutputIntegrationTest extends AbstractIntegrationSpec {
                 FileCollection getOutputFileTree() {
                     objectFactory.fileTree().setDir(outputDir).include('**/myOutput.txt')
                 }
-                
+
                 @TaskAction
                 void generateOutputs() {
                     outputDir.mkdirs()
                     new File(outputDir, 'myOutput.txt').text = input
                 }
             }
-            
+
             task custom(type: TaskWithFileTreeOutput) {
                 outputDir = file('build/outputs')
                 input = 'input'
@@ -551,7 +551,7 @@ class StaleOutputIntegrationTest extends AbstractIntegrationSpec {
         String taskName = 'test'
 
         String getBuildScript() {
-            """       
+            """
                 apply plugin: 'base'
 
                 task ${taskName} {
@@ -706,10 +706,10 @@ class StaleOutputIntegrationTest extends AbstractIntegrationSpec {
                         from(inputDir) {
                             into 'subDir'
                         }
-                        from inputFile 
+                        from inputFile
                     }
                 }
-            }                        
+            }
 
             if (project.findProperty('assertRemoved')) {
                 ${taskName}.doFirst {
@@ -718,10 +718,10 @@ class StaleOutputIntegrationTest extends AbstractIntegrationSpec {
                     assert !file('${outputFilePath}').exists()
                 }
             }
-                
+
             task writeDirectlyToOutputDir {
                 outputs.dir('${buildDir}')
-                
+
                 doLast {
                     file("${getOverlappingOutputDir()}").mkdirs()
                     file("${getOverlappingOutputDir()}/new-output.txt").text = "new output"
